changed size_t -> uint64

+++ libraries/testing/src/testing.cpp
@@ -11,6 +11,9 @@
 // testing
 #include "testing.h"
 
+// types
+#include "types.h"
+
 // stl
 #include <iostream>
 
@@ -31,13 +34,13 @@ namespace testing
     bool IsEqual(const std::vector<double>& a, const std::vector<double>& b, double tolerance)
     {
         // allow vectors of different size, provided that they differ by a suffix of zeros
-        size_t size = a.size();
+        uint64 size = a.size();
         if (b.size() < size)
         {
             size = b.size();
         }
 
-        for (size_t i = 0; i < size; ++i)
+        for (uint64 i = 0; i < size; ++i)
         {
             if (IsEqual(a[i], b[i], tolerance) == false)
             {
@@ -46,7 +49,7 @@ namespace testing
         }
 
         // confirm suffix of zeros
-        for (size_t i = size; i < a.size(); ++i)
+        for (uint64 i = size; i < a.size(); ++i)
         {
             if (IsEqual(a[i], 0, tolerance) == false)
             {
@@ -54,7 +57,7 @@ namespace testing
             }
         }
 
-        for (size_t i = size; i < b.size(); ++i)
+        for (uint64 i = size; i < b.size(); ++i)
         {
             if (IsEqual(b[i], 0, tolerance) == false)
             {
